Wrap enum for config

+++ chains/solana/contracts/programs/ccip-offramp/src/instructions/v1/execute.rs
@@ -41,7 +41,7 @@ impl Execute for Impl {
                 &config.ocr3[OcrPluginType::Execution as usize],
                 &ctx.accounts.sysvar_instructions,
                 ctx.accounts.authority.key(),
-                OcrPluginType::Execution as u8,
+                OcrPluginType::Execution,
                 report_context,
                 &Ocr3ReportForExecutionReportSingleChain(&execution_report),
                 Signatures {
